fix: CalculatorProcessorInterface를 제거하고 CalculatorProcessor 클래스만 사용한다.
